ecs-settings-applier: conditional compilation
This commit will compile ecs-settings-applier with a no-op body when any
variant other than aws-ecs-1 is used.  This allows for the top-level
`cargo make` targets like `unit-tests` to function properly for all
variants.
