+++ sources/api/ecs-settings-applier/build.rs
@@ -6,13 +6,18 @@ use std::io::Write;
 use std::path::PathBuf;
 
 fn main() {
+    println!("cargo:rerun-if-env-changed=VARIANT");
+    if let Ok(variant) = env::var("VARIANT") {
+        println!("cargo:rustc-cfg=variant=\"{}\"", variant);
+    }
+
     // Check for environment variable "SKIP_README". If it is set,
     // skip README generation
     if env::var_os("SKIP_README").is_some() {
         return;
     }
 
-    let mut source = File::open("src/main.rs").unwrap();
+    let mut source = File::open("src/ecs.rs").unwrap();
     let mut template = File::open("README.tpl").unwrap();
 
     let content = cargo_readme::generate_readme(
